+++ src/main/java/com/cognixia/jump/model/SavingsAccount.java
@@ -4,10 +4,10 @@ public class SavingsAccount extends Account{
 
 	private Integer id;
 	private String type = "Savings";
-	private Double balance;
+	private Float balance;
 	private Customer customer;
 	
-	public SavingsAccount(Double balance, Customer customer) {
+	public SavingsAccount(Float balance, Customer customer) {
 		super();
 		this.id = super.getId();
 		this.balance = balance;
@@ -16,12 +16,12 @@ public class SavingsAccount extends Account{
 
 	
 	@Override
-	public void deposit(Double amount) {
+	public void deposit(Float amount) {
 		this.balance += amount;
 	}
 
 	@Override
-	public void withdraw(Double amount) {
+	public void withdraw(Float amount) {
 		this.balance -= amount;
 	}
 
@@ -34,7 +34,7 @@ public class SavingsAccount extends Account{
 		this.id = id;
 	}
 
-	public Double getBalance() {
+	public Float getBalance() {
 		return balance;
 	}
